CASM-3632: Move some common script functions into shared library
(cherry picked from commit cf73ea41da357da8d886544381c205f85a72a913)
